# api/schemas.py
from pydantic import BaseModel
from typing import Optional, List, Dict, Any
import datetime

# --- Tag Schemas ---
class TagBase(BaseModel):
    name: str

class TagCreate(TagBase):
    pass

class Tag(TagBase):
    id: int

    class Config:
        from_attributes = True

# --- Album Schemas ---
class AlbumBase(BaseModel):
    name: str
    description: Optional[str] = None

class AlbumCreate(AlbumBase):
    pass

class Album(AlbumBase):
    id: int
    images: List["Image"] = []

    class Config:
        from_attributes = True


# --- Location Schemas ---
class LocationBase(BaseModel):
    latitude: float
    longitude: float
    address: Optional[str] = None

class LocationCreate(LocationBase):
    pass

class Location(LocationBase):
    id: int

    class Config:
        from_attributes = True

# --- Metadata Schemas ---
class MetadataBase(BaseModel):
    camera_model: Optional[str] = None
    date_taken: Optional[datetime.datetime] = None
    f_number: Optional[float] = None
    exposure_time: Optional[str] = None
    iso: Optional[int] = None
    focal_length: Optional[str] = None
    lens_model: Optional[str] = None
    raw_exif: Optional[Dict[str, Any]] = None

class MetadataCreate(MetadataBase):
    pass

class Metadata(MetadataBase):
    id: int
    location: Optional[Location] = None

    class Config:
        from_attributes = True

# --- Image Schemas ---
class ImageBase(BaseModel):
    filename: str
    filepath: str
    upload_date: datetime.datetime
    resolution: str
    image_size: int
    mimetype: Optional[str] = None


class ImageCreate(ImageBase):
    pass


class ImageUpdate(BaseModel):
    filepath: Optional[str] = None
    resolution: Optional[str] = None
    image_size: Optional[int] = None
    mimetype: Optional[str] = None
    status: Optional[str] = None

# This is the main schema for reading an image, includes nested data
class Image(ImageBase):
    id: int
    details: Optional[Metadata] = None
    tags: List[Tag] = []
    albums: List[AlbumBase] = []
    is_favorite: bool = False
    thumbnail_url: Optional[str] = None
    medium_url: Optional[str] = None
    large_url: Optional[str] = None
    status: str = "processing"

    class Config:
        from_attributes = True

Album.model_rebuild()

# --- Search Schemas ---
class SearchQuery(BaseModel):
    """
    A structured representation of a user's natural language search query.
    The AI model will populate the fields of this model based on the user's input.
    """
    tags: Optional[List[str]] = None
    location: Optional[str] = None
    camera_model: Optional[str] = None
    date_query: Optional[str] = None