# openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /api/v1/images:
    post:
      summary: Upload Image
      description: 'Upload and process a new image file.

        Returns 409 if file exists, 400 if invalid file type, 500 for processing errors.'
      operationId: upload_image_api_v1_images_post
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_image_api_v1_images_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Read Images
      operationId: read_images_api_v1_images_get
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Image'
                title: Response Read Images Api V1 Images Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Body_upload_image_api_v1_images_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_upload_image_api_v1_images_post
    Image:
      properties:
        filename:
          type: string
          title: Filename
        filepath:
          type: string
          title: Filepath
        upload_date:
          type: string
          format: date-time
          title: Upload Date
        resolution:
          type: string
          title: Resolution
        image_size:
          type: integer
          title: Image Size
        mimetype:
          anyOf:
          - type: string
          - type: 'null'
          title: Mimetype
        id:
          type: integer
          title: Id
        details:
          anyOf:
          - $ref: '#/components/schemas/Metadata'
          - type: 'null'
        tags:
          items:
            $ref: '#/components/schemas/Tag'
          type: array
          title: Tags
          default: []
        is_favorite:
          type: boolean
          title: Is Favorite
          default: false
        thumbnail_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Thumbnail Url
        medium_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Medium Url
        large_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Large Url
      type: object
      required:
      - filename
      - filepath
      - upload_date
      - resolution
      - image_size
      - id
      title: Image
    Location:
      properties:
        latitude:
          type: number
          title: Latitude
        longitude:
          type: number
          title: Longitude
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        id:
          type: integer
          title: Id
      type: object
      required:
      - latitude
      - longitude
      - id
      title: Location
    Metadata:
      properties:
        camera_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Camera Model
        date_taken:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date Taken
        f_number:
          anyOf:
          - type: number
          - type: 'null'
          title: F Number
        exposure_time:
          anyOf:
          - type: string
          - type: 'null'
          title: Exposure Time
        iso:
          anyOf:
          - type: integer
          - type: 'null'
          title: Iso
        focal_length:
          anyOf:
          - type: string
          - type: 'null'
          title: Focal Length
        lens_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Lens Model
        raw_exif:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Raw Exif
        id:
          type: integer
          title: Id
        location:
          anyOf:
          - $ref: '#/components/schemas/Location'
          - type: 'null'
      type: object
      required:
      - id
      title: Metadata
    Tag:
      properties:
        name:
          type: string
          title: Name
        id:
          type: integer
          title: Id
      type: object
      required:
      - name
      - id
      title: Tag